Crate spin_factor_key_value

Re-exports

pub use runtime_config::RuntimeConfig;

Modules

runtime_config

Structs

AppState

CachingStoreManager

Wrap each Store produced by the inner StoreManager in an asynchronous, write-behind cache.

DelegatingStoreManager

A StoreManager which delegates to other StoreManagers based on the store label.

InstanceBuilder

KeyValueDispatch

KeyValueFactor

A factor that provides key-value storage.

Enums

Error

The set of errors which may be raised by functions in this interface

SwapError

SwapError are errors that occur during compare and swap operations

Constants

KEY_VALUE_STORES_KEY

Metadata key for key-value stores.

Traits

Cas

Cas trait describes the interface a key value compare and swap implementor must fulfill.

Store

StoreManager

Functions

log_cas_error

log_error