spin_factor_key_value::runtime_config::spin
Trait MakeKeyValueStore

pub trait MakeKeyValueStore:
    'static
    + Send
    + Sync {
    type RuntimeConfig: DeserializeOwned;
    type StoreManager: StoreManager;

    const RUNTIME_CONFIG_TYPE: &'static str;

    // Required method
    fn make_store(
        &self,
        runtime_config: Self::RuntimeConfig,
    ) -> Result<Self::StoreManager>;
}

Defines the construction of a key value store from a serialized runtime config.

Required Associated Constants

const RUNTIME_CONFIG_TYPE: &'static str

Unique type identifier for the store.

Required Associated Types

type RuntimeConfig: DeserializeOwned

Runtime configuration for the store.

type StoreManager: StoreManager

The store manager for the store.

Required Methods

fn make_store( &self, runtime_config: Self::RuntimeConfig, ) -> Result<Self::StoreManager>

Creates a new store manager from the runtime configuration.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.